[SPARK-54739][INFRA] Cancel workflow if a new commit is pushed
### What changes were proposed in this pull request?

We cancel the workflow if there is new commit pushed to the same branch.

### Why are the changes needed?

The existing workflow is super long (1~2hours). github limits the concurrent jobs for each repo, which means if we push a few commits in a short period of time, the latest commit will have to wait for the old ones to finish. However, in reality, we don't care about the result of the previous commits - they are irrelevant.

### Does this PR introduce _any_ user-facing change?

No, it's an infra change.

### How was this patch tested?

It's not code change.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53008 from gaogaotiantian/cancel-same-pr-workflow.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>