[SPARK-54561][PYTHON] Support breakpoint() for `run-tests.py`
### What changes were proposed in this pull request?

We hijack the stdin/stdout/stderr in `run-tests.py` so we can support `breakpoint()` in test cases. The users can now write a `breakpoint()` in a test and debug the program when they use `run-tests.py`.

### Why are the changes needed?

It's difficult to debug a failed test. This also lays the foundation to temporarily redirect stderr/stdout to sys.stdout.

### Does this PR introduce _any_ user-facing change?

No. The original behavior should be preserved. The performance impact is negligible.

### How was this patch tested?

Add a line of `breakpoint()` in the test case then run the test with `./run-tests --testname="..."`. The code will hit the breakpoint and bring up pdb. The user can debug the test using pdb interface.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53273 from gaogaotiantian/support-breakpoint.

Lead-authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Co-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>