[SPARK-54572][PYTHON][FOLLOWUP] Cleanup all child processes for vscode script
### What changes were proposed in this pull request?

Cleanup all the child processes when `run-with-vscode-breakpoint` exits with `kill -- -$$` (kill the process group of this script).

### Why are the changes needed?

`debugpy` will leave some long live processes (probably due to how we used it, multiple connections to vscode). Users might not aware of this or they don't know how to clear them. This is a helpful way to just kill everything started by the script. Considering the script is for dev-debug only, it should be safe.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Locally confirmed that all the debug processes are killed after script exits.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53551 from gaogaotiantian/cleanup-process.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>