[SPARK-53991][SQL][TEST][FOLLOWUP] Make KLL quantile golden file tests deterministic
### What changes were proposed in this pull request?

In https://github.com/apache/spark/pull/52800, we added SQL support for KLL quantiles functions based on DataSketches.

In this PR, we update some of the golden file tests to make them deterministic.

### Why are the changes needed?

The previous tests generated string summaries of KLL quantile sketches and then split them by newlines and made case-sensitive checks for substrings. It turns out this was brittle, so this PR updates the tests to avoid the newline-splitting and makes the substring checks case-insenstiive.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

This PR updates test coverage only.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53549 from dtenedor/kll-quantile-golden-files-fix.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>