[SPARK-54735][SQL] Properly preserve column comments in view with SCHEMA EVOLUTION
### What changes were proposed in this pull request?
In this PR I propose a fix to avoid overwriting user defined comments in views that have schema evolution enabled.

### Why are the changes needed?
When user creates a view with `SCHEMA EVOLUTION`, and then changes the comments on view columns, by using `DESCRIBE`, they are able to see new comments, but after executing a query on a view, comment values are reversed to a values from the underlying table (schema evolution enabled).

In the proposed fix, if the column name is changed, comments from the table will be adopted (consistent with evolution semantics).

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Unit test in `SQLViewTestSuite`.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #53508 from miland-db/milan-dankovic_data/master.

Authored-by: Milan Dankovic <milan.dankovic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>