[SPARK-54800] Changed default implementation for isObjectNotFoundException
### What changes were proposed in this pull request?

Changed that `isObjectNotFoundException`default implementation considers an exception when its `SQLState` starts with 42.

### Why are the changes needed?

Before, if `isObjectNotFound` was not overridden (for example, in a custom Dialect), then when any other error was thrown that was not ‘table does not exist’, it would still be reported as a TableNotFound error, which is incorrect

### Does this PR introduce _any_ user-facing change?

Updated method in JDBCDialects.

### How was this patch tested?

In another PR #53541.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53456 from SavicStefan/stefan-savic_data/stefan.

Authored-by: Stefan Savić <stefan.savic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit aad5b6e19c6eb9280ae7e1f04cd39b2f055aa3ce)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>