[SPARK-54799][SQL] Refactor `UnpivotCoercion`
### What changes were proposed in this pull request?
In this PR I propose to refactor `UnpivotCoercion`.

### Why are the changes needed?
It's needed to make it reusable in the single-pass resolver. In addition to this change, we add timezone to the `Cast` in the rule in place to avoid traversing the tree later in the single-pass and thus make it slower.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #53564 from mihailoale-db/refactorunpivotcoercion.

Authored-by: mihailoale-db <mihailo.aleksic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>