[SPARK-54844][BUILD] Upgrade maven.version to 3.9.12
### What changes were proposed in this pull request?

This PR upgrades the Maven version from 3.9.11 to 3.9.12 in the project's root POM file.

### Why are the changes needed?

Witness downloading issue in GHA

### Does this PR introduce _any_ user-facing change?

No. This is a build-only change that does not affect runtime behavior or user-facing features.

### How was this patch tested?

This change only updates the Maven version property in the POM file. The existing CI/CD pipelines will validate the build with the new Maven version.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53610 from yaooqinn/mvn.

Authored-by: Kent Yao <kentyao@microsoft.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>