[SPARK-54834][SQL] Add new interfaces SimpleProcedure and SimpleFunction
### What changes were proposed in this pull request?

It's common that a procedure does not have any overload and the `bind` method always return the same `BoundProcedure`. This PR adds a new interface `SimpleProcedure ` for this use case, which allows people to implement a `BoundProcedure` directly without thinking about how to bind. The same applies to v2 functions.

### Why are the changes needed?

Simplify a common use case when people implement procedures and functions.

### Does this PR introduce _any_ user-facing change?

No, it's developer facing

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

cursor 2.2.43

Closes #53595 from cloud-fan/procedure.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>